requirements.txt
Django~=3.2.12
gunicorn==20.0.4
djangorestframework==3.12.2
dataclasses==0.6
pytest-django==4.0.0
faker==4.1.2
